Rename sci to cri
