+++ rs/src/ir.rs
@@ -53,7 +53,7 @@ pub enum Operator {
     LessThanOrEqual,
     Plus,
     ReadRtCol(crate::fql::backend::lmi::ColId), /* TODO re-design declarations for operator body */
-    ReadRtColSci(crate::fql::backend::sci::ColId), /* TODO re-design declarations for operator body */
+    ReadRtColSci(crate::fql::backend::cri::ColId), /* TODO re-design declarations for operator body */
 }
 
 /* LetExpr */
